[ADD] 진행 중인 대결 상세 화면 API 적용

+++ domain/src/main/java/kr/co/domain/model/AllBattleCompletionDto.kt
@@ -1,7 +1,7 @@
 package kr.co.domain.model
 
 data class AllBattleCompletionDto(
-    val battleId: Int,
+    val battleId: Long,
     val hasResult: Int,
     val otherProfileImg: String,
     val winnerId: Int,
